add delete producers
