+++ rubix-ui/frontend/src/components/hosts/host/flow/flowNetworks/producers/views/table.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { Space, Spin } from "antd";
 import { main, model } from "../../../../../../../../wailsjs/go/models";
@@ -7,6 +7,7 @@ import { PRODUCER_HEADERS } from "../../../../../../../constants/headers";
 import RbTable from "../../../../../../../common/rb-table";
 import {
   RbAddButton,
+  RbDeleteButton,
   RbRefreshButton,
 } from "../../../../../../../common/rb-table-actions";
 import { CreateEditModal } from "./create";
@@ -15,11 +16,12 @@ import UUIDs = main.UUIDs;
 import Producer = model.Producer;
 
 export const ProducersTable = (props: any) => {
-  const { data, isFetching, refreshList } = props;
   let { connUUID = "", hostUUID = "" } = useParams();
   const [selectedUUIDs, setSelectedUUIDs] = useState([] as Array<UUIDs>);
+  const [producers, setProducers] = useState([] as Producer[]);
   const [currentItem, setCurrentItem] = useState({} as Producer);
   const [isModalVisible, setIsModalVisible] = useState(false);
+  const [isFetching, setIsFetching] = useState(false);
 
   let factory = new FlowProducerFactory();
   factory.connectionUUID = connUUID;
@@ -61,21 +63,43 @@ export const ProducersTable = (props: any) => {
     setCurrentItem({} as Producer);
   };
 
+  const fetch = async () => {
+    try {
+      setIsFetching(true);
+      const res = await factory.GetAll();
+      setProducers(res);
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setIsFetching(false);
+    }
+  };
+
+  const bulkDelete = async () => {
+    await factory.BulkDelete(selectedUUIDs);
+    fetch();
+  };
+
+  useEffect(() => {
+    fetch();
+  }, []);
+
   return (
     <>
-      <RbRefreshButton refreshList={refreshList} />
+      <RbRefreshButton refreshList={fetch} />
+      <RbDeleteButton bulkDelete={bulkDelete} />
       <RbAddButton showModal={() => showModal({} as Producer)} />
       <RbTable
         rowKey="uuid"
         rowSelection={rowSelection}
-        dataSource={data}
+        dataSource={producers}
         columns={columns}
         loading={{ indicator: <Spin />, spinning: isFetching }}
       />
       <CreateEditModal
         currentItem={currentItem}
         isModalVisible={isModalVisible}
-        refreshList={refreshList}
+        refreshList={fetch}
         onCloseModal={onCloseModal}
       />
     </>
